Bugfix/bugfixes (#75)
* Moved Search to Component

* Moved UserProfile to Component

* Fixed UserPreview Length

* Fixed Long Video Title in Preivew

* Added Verification Page

* Fix
