+++ src/Common/Helper/UserNameShorter.tsx
@@ -1,6 +1,6 @@
-export default function shortCreatorName(creatorName: string) {
-    if (creatorName.length > 10){
-        creatorName = creatorName.slice(0, 10) + "..."
+export default function shortCreatorName(creatorName: string, length: number) {
+    if (creatorName.length > length){
+        creatorName = creatorName.slice(0, length) + "..."
     } 
     return creatorName;
 }
