+++ src/Common/UserPreview/UserPreview.css
@@ -8,7 +8,8 @@
     background-color: white;
     justify-content: center;
     border-radius: 10px;
-    width: fit-content;
+    width: max-content;
+    min-width: 94%;
     padding-right: 1em;
 }
